Sahayak/requirements.txt
streamlit
openai
gTTS
python-dotenv